feat: 서버 컴포넌트 활용을 위한 대시보드 분리 및 서버 prefetch 로직 생성
